Portage Windows ++ (deleted blocAir, improved fluidity, errors ...)

+++ MyCCraft/Fun.cpp
@@ -25,9 +25,9 @@ public:
                 for(int y=0; y< 256; y++){
                     bloc* aBloc;
                     if(y<height){
-                        aBloc = new bloc(1, x, y, z, 1.0);
+                        aBloc = new bloc(1, (float)x, (float)y, (float)z, 1.0f);
                     }else{
-                        aBloc = new air(0, x, y, z, 0.0);
+                        aBloc = new air(0, (float)x, (float)y, (float)z, 0.0f);
                         aBloc->visible = false;
                     }
                     if(y<height-1)
@@ -42,20 +42,20 @@ public:
         
         
         for(int cpt=0; cpt<10; cpt++){
-            aChunk->setBloc(new bloc(13.0, height+cpt, 15.0-cpt, 0.0, 0.6, 0.2, (float)cpt/10.0, 1.0));
-            aChunk->setBloc(new bloc(14.0, height+cpt, 15.0-cpt, 0.0, 0.6, 0.2, (float)cpt/10.0, 1.0));
-            aChunk->setBloc(new bloc(15.0, height+cpt, 15.0-cpt, 0.0, 0.6, 0.2, (float)cpt/10.0, 1.0));
+            aChunk->setBloc(new bloc(13.0f, (float)height+(float)cpt, (float)15.0f-(float)cpt, 0.0f, 0.6f, 0.2f, (float)cpt/10.0f, 1.0f));
+            aChunk->setBloc(new bloc(14.0f, (float)height+(float)cpt, (float)15.0f-(float)cpt, 0.0f, 0.6f, 0.2f, (float)cpt/10.0f, 1.0f));
+            aChunk->setBloc(new bloc(15.0f, (float)height+(float)cpt, (float)15.0f-(float)cpt, 0.0f, 0.6f, 0.2f, (float)cpt/10.0f, 1.0f));
         }
         
         
         for(int cpt=0; cpt < 10; cpt++){
-            aChunk->setBloc(new bloc(cpt+5.0, height, 0.0, 0.0, 0.6, (float)cpt/10.0, 0.2, 1.0, 1.0));
-            aChunk->setBloc(new bloc(cpt+5.0, height+1, 0.0, 0.0, 0.6, (float)cpt/10.0, 0.2, 1.0, 1.0));
-            aChunk->setBloc(new bloc(cpt+5.0, height+2, 0.0, 0.0, 0.6, (float)cpt/10.0, 0.2, 1.0, 1.0));
-            aChunk->setBloc(new bloc(cpt+5.0, height+2, 1.0, 0.0, 0.6, (float)cpt/10.0, 0.2, 1.0, 1.0));
-            aChunk->setBloc(new bloc(cpt+5.0, height+2, 2.0, 0.0, 0.6, (float)cpt/10.0, 0.2, 1.0, 1.0));
-            aChunk->setBloc(new bloc(cpt+5.0, height+1, 2.0, 0.0, 0.6, (float)cpt/10.0, 0.2, 1.0, 1.0));
-            aChunk->setBloc(new bloc(cpt+5.0, height, 2.0, 0.0, 0.6, (float)cpt/10.0, 0.2, 1.0, 1.0));
+            aChunk->setBloc(new bloc((float)cpt+5.0f, (float)height, 0.0f, 0.0f, 0.6f, (float)cpt/10.0f, 0.2f, 1.0f, 1.0f));
+            aChunk->setBloc(new bloc((float)cpt+5.0f, (float)height+1.0f, 0.0f, 0.0f, 0.6f, (float)cpt/10.0f, 0.2f, 1.0f, 1.0f));
+            aChunk->setBloc(new bloc((float)cpt+5.0f, (float)height+2.0f, 0.0f, 0.0f, 0.6f, (float)cpt/10.0f, 0.2f, 1.0f, 1.0f));
+            aChunk->setBloc(new bloc((float)cpt+5.0f, (float)height+2.0f, 1.0f, 0.0f, 0.6f, (float)cpt/10.0f, 0.2f, 1.0f, 1.0f));
+            aChunk->setBloc(new bloc((float)cpt+5.0f, (float)height+2.0f, 2.0f, 0.0f, 0.6f, (float)cpt/10.0f, 0.2f, 1.0f, 1.0f));
+            aChunk->setBloc(new bloc((float)cpt+5.0f, (float)height+1.0f, 2.0f, 0.0f, 0.6f, (float)cpt/10.0f, 0.2f, 1.0f, 1.0f));
+            aChunk->setBloc(new bloc((float)cpt+5.0f, (float)height, 2.0f, 0.0f, 0.6f, (float)cpt/10.0f, 0.2f, 1.0f, 1.0f));
         }
