+++ MyCCraft/Mountain.cpp
@@ -25,9 +25,9 @@ public:
                 for(int y=0; y< 256; y++){
                     bloc* aBloc;
                     if(y<height){
-                        aBloc = new bloc(1, x, y, z, 1.0);
+                        aBloc = new bloc(1, (float)x, (float)y, (float)z, 1.0);
                     }else{
-                        aBloc = new air(0, x, y, z, 0.0);
+                        aBloc = new air(0, (float)x, (float)y, (float)z, 0.0);
                         aBloc->visible = false;
                     }
                     if(y<height-1)
@@ -43,33 +43,33 @@ public:
         
         for(int cpt=1; cpt<8; cpt++){
             for(int cpt2=1; cpt2<8; cpt2++){
-                aChunk->setBloc(new bloc(cpt, height+cpt+cpt2-10, cpt2, 0.0, 0.6, (float)cpt2/10.0, (float)cpt/10.0, 1.0));
-                aChunk->setBloc(new bloc(cpt, height+cpt+cpt2-10, cpt2, 0.0, 0.6, (float)cpt2/10.0, (float)cpt/10.0, 1.0));
-                aChunk->setBloc(new bloc(cpt, height+cpt+cpt2-10, cpt2, 0.0, 0.6, (float)cpt2/10.0, (float)cpt/10.0, 1.0));
+                aChunk->setBloc(new bloc((float)cpt, (float)height+cpt+cpt2-10, (float)cpt2, 0.0f, 0.6f, (float)cpt2/10.0f, (float)cpt/10.0f, 1.0f));
+                aChunk->setBloc(new bloc((float)cpt, (float)height+cpt+cpt2-10, (float)cpt2, 0.0f, 0.6f, (float)cpt2/10.0f, (float)cpt/10.0f, 1.0f));
+                aChunk->setBloc(new bloc((float)cpt, (float)height+cpt+cpt2-10, (float)cpt2, 0.0f, 0.6f, (float)cpt2/10.0f, (float)cpt/10.0f, 1.0f));
             }
         }
         
         for(int cpt=14; cpt>=8; cpt--){
             for(int cpt2=14; cpt2>=8; cpt2--){
-                aChunk->setBloc(new bloc(cpt, 40-height-cpt-cpt2, cpt2, 0.0, 0.6, (float)cpt2/10.0, (float)cpt/10.0, 1.0));
-                aChunk->setBloc(new bloc(cpt, 40-height-cpt-cpt2, cpt2, 0.0, 0.6, (float)cpt2/10.0, (float)cpt/10.0, 1.0));
-                aChunk->setBloc(new bloc(cpt, 40-height-cpt-cpt2, cpt2, 0.0, 0.6, (float)cpt2/10.0, (float)cpt/10.0, 1.0));
+                aChunk->setBloc(new bloc((float)cpt, (float)40-height-cpt-cpt2, (float)cpt2, 0.0f, 0.6f, (float)cpt2/10.0f, (float)cpt/10.0f, 1.0f));
+                aChunk->setBloc(new bloc((float)cpt, (float)40-height-cpt-cpt2, (float)cpt2, 0.0f, 0.6f, (float)cpt2/10.0f, (float)cpt/10.0f, 1.0f));
+                aChunk->setBloc(new bloc((float)cpt, (float)40-height-cpt-cpt2, (float)cpt2, 0.0f, 0.6f, (float)cpt2/10.0f, (float)cpt/10.0f, 1.0f));
             }
         }
         
         for(int cpt=14; cpt>=8; cpt--){
             for(int cpt2=1; cpt2<8; cpt2++){
-                aChunk->setBloc(new bloc(cpt, 25-height-cpt+cpt2, cpt2, 0.0, 0.6, (float)cpt2/10.0, (float)cpt/10.0, 1.0));
-                aChunk->setBloc(new bloc(cpt, 25-height-cpt+cpt2, cpt2, 0.0, 0.6, (float)cpt2/10.0, (float)cpt/10.0, 1.0));
-                aChunk->setBloc(new bloc(cpt, 25-height-cpt+cpt2, cpt2, 0.0, 0.6, (float)cpt2/10.0, (float)cpt/10.0, 1.0));
+                aChunk->setBloc(new bloc((float)cpt, (float)25-height-cpt+cpt2, (float)cpt2, 0.0f, 0.6f, (float)cpt2/10.0f, (float)cpt/10.0f, 1.0f));
+                aChunk->setBloc(new bloc((float)cpt, (float)25-height-cpt+cpt2, (float)cpt2, 0.0f, 0.6f, (float)cpt2/10.0f, (float)cpt/10.0f, 1.0f));
+                aChunk->setBloc(new bloc((float)cpt, (float)25-height-cpt+cpt2, (float)cpt2, 0.0f, 0.6f, (float)cpt2/10.0f, (float)cpt/10.0f, 1.0f));
             }
         }
         
         for(int cpt=1; cpt<8; cpt++){
             for(int cpt2=14; cpt2>=8; cpt2--){
-                aChunk->setBloc(new bloc(cpt, 25-height+cpt-cpt2, cpt2, 0.0, 0.6, (float)cpt2/10.0, (float)cpt/10.0, 1.0));
-                aChunk->setBloc(new bloc(cpt, 25-height+cpt-cpt2, cpt2, 0.0, 0.6, (float)cpt2/10.0, (float)cpt/10.0, 1.0));
-                aChunk->setBloc(new bloc(cpt, 25-height+cpt-cpt2, cpt2, 0.0, 0.6, (float)cpt2/10.0, (float)cpt/10.0, 1.0));
+                aChunk->setBloc(new bloc((float)cpt, (float)25-height+cpt-cpt2, (float)cpt2, 0.0f, 0.6f, (float)cpt2/10.0f, (float)cpt/10.0f, 1.0f));
+                aChunk->setBloc(new bloc((float)cpt, (float)25-height+cpt-cpt2, (float)cpt2, 0.0f, 0.6f, (float)cpt2/10.0f, (float)cpt/10.0f, 1.0f));
+                aChunk->setBloc(new bloc((float)cpt, (float)25-height+cpt-cpt2, (float)cpt2, 0.0f, 0.6f, (float)cpt2/10.0f, (float)cpt/10.0f, 1.0f));
             }
         }
